fix(name): 优化代码
提取badge的js，删除修改时的外部API，增加点人姓名时的跳转。
