+++ src/main/resources/static/js/admin/expense-report-add.js
@@ -1,24 +1,5 @@
 $(function () {
     let $detail_body = $('#detail-body')
-    $.ajax({
-        url: '/expenseReport/getViewList',
-        type: 'POST',
-        async: false,
-        cache: false,
-        dataType: 'json',
-        contentType: 'application/json;charset=utf-8',
-        data: JSON.stringify({}),
-        success: data => {
-            if (data.success) {
-                let len = data.data.records.length
-                if (0 !== len) {
-                    $('.badge').text(len)
-                }
-            } else {
-                lightyear.notify(data.errMsg, 'danger', 2000, 'mdi mdi-emoticon-sad', 'top', 'center')
-            }
-        }
-    })
 
     $detail_body.on('change', '.detail_image', async function () {
         lightyear.loading('show');
