+++ src/main/resources/static/js/admin/expense-report-view.js
@@ -28,26 +28,6 @@ $(function () {
 
     getList(request_condition)
 
-    $.ajax({
-        url: '/expenseReport/getViewList',
-        type: 'POST',
-        async: false,
-        cache: false,
-        dataType: 'json',
-        contentType: 'application/json;charset=utf-8',
-        data: JSON.stringify({}),
-        success: function (data) {
-            if (data.success) {
-                let len = data.data.records.length
-                if (0 !== len) {
-                    $('.badge').text(len)
-                }
-            } else {
-                lightyear.notify(data.errMsg, 'danger', 2000, 'mdi mdi-emoticon-sad', 'top', 'center')
-            }
-        }
-    })
-
     function getList(data) {
         $.ajax({
             url: '/expenseReport/getViewList',
@@ -92,7 +72,7 @@ $(function () {
         data.map(function (item, index) {
             html += '<tr>'
                 + '<td>' + (index + 1) + '</td>'
-                + '<td>' + item.createEmployee.name + '</td>'
+                + '<td><a class="my-link" href="/employee/goEmployee?emId=' + (item.createEmployee.emId) + '" title="查看" data-toggle="tooltip">' + item.createEmployee.name + '</a></td>'
                 + '<td>' + item.createEmployee.department.name + '</td>'
                 + '<td>' + item.cause + '</td>'
                 + '<td>' + item.createTime + '</td>'
