+++ src/main/resources/templates/authorities.html
@@ -418,6 +418,7 @@
 <script type="text/javascript" src="../static/js/main.min.js"></script>
 
 <script type="text/javascript" src="../static/js/chosen.jquery.min.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript">
     $(function(){
         //动态选择框，上下级选中状态变化
