+++ src/main/resources/templates/department-list.html
@@ -97,6 +97,7 @@
 <script type="text/javascript" src="../static/js/jquery.pagination.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
 <script type="text/javascript" src="../static/js/admin/common.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/department-list.js"></script>
 <script type="text/javascript">
 </script>
