+++ src/main/resources/templates/editPwdPage.html
@@ -66,6 +66,7 @@
 <script type="text/javascript" src="../static/js/jconfirm/jquery-confirm.min.js"></script>
 <script type="text/javascript" src="../static/js/main.min.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/editPassword.js"></script>
 <script type="text/javascript">
 </script>
