+++ src/main/resources/templates/employee-edit.html
@@ -85,6 +85,7 @@
 <script type="text/javascript" src="../static/js/bootstrap-select/bootstrap-select.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
 <script type="text/javascript" src="../static/js/admin/common.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/employee-operation.js"></script>
 </body>
 </html>
