+++ src/main/resources/templates/employee-list.html
@@ -104,6 +104,7 @@
 <script type="text/javascript" src="../static/js/jquery.pagination.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
 <script type="text/javascript" src="../static/js/admin/common.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/employee-list.js"></script>
 <script type="text/javascript" src="../static/js/bootstrap-select/bootstrap-select.js"></script>
 </body>
