+++ src/main/resources/templates/expense-report-add.html
@@ -172,6 +172,7 @@
 <script type="text/javascript" src="../static/js/admin/common.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
 <script type="text/javascript" src="../static/js/admin/date-format.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/expense-report-add.js"></script>
 </body>
 </html>
