+++ src/main/resources/templates/main.html
@@ -9,6 +9,7 @@
     <link href="../static/css/materialdesignicons.min.css" rel="stylesheet">
     <link href="../static/css/style.min.css" rel="stylesheet">
     <link href="../static/css/pagination.css" rel="stylesheet">
+    <link href="../static/css/my-link.css" rel="stylesheet">
 </head>
 
 <body>
@@ -162,6 +163,7 @@
 <script type="text/javascript" src="../static/js/jquery.pagination.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
 <script type="text/javascript" src="../static/js/Chart.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/mainPage.js"></script>
 </body>
 </html>
