+++ src/main/resources/templates/position-add.html
@@ -72,6 +72,7 @@
 <script type="text/javascript" src="../static/js/main.min.js"></script>
 <script type="text/javascript" src="../static/js/admin/common.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/position-operation.js"></script>
 </body>
 </html>
