+++ src/main/resources/templates/position-list.html
@@ -87,6 +87,7 @@
 <script type="text/javascript" src="../static/js/jquery.pagination.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
 <script type="text/javascript" src="../static/js/admin/common.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/position-list.js"></script>
 <script type="text/javascript">
 </script>
