+++ src/main/resources/templates/view-expense-report-detail.html
@@ -13,6 +13,7 @@
     <link href="../static/css/style.min.css" rel="stylesheet">
     <link href="../static/css/input-file.css" rel="stylesheet">
     <link href="../static/css/truncate.css" rel="stylesheet">
+    <link href="../static/css/my-link.css" rel="stylesheet">
 </head>
 
 <body>
@@ -111,6 +112,7 @@
 <script type="text/javascript" src="../static/js/jconfirm/jquery-confirm.min.js"></script>
 <script type="text/javascript" src="../static/js/admin/common.js"></script>
 <script type="text/javascript" src="../static/js/admin/log-out.js"></script>
+<script type="text/javascript" src="../static/js/admin/getBadge.js"></script>
 <script type="text/javascript" src="../static/js/admin/view-expense-report-detail.js"></script>
 </body>
 </html>
